// src/app/shared/redux/project/project.type.ts
export enum ProjectType {
    LIST = '[PROJECT API] List',
    LIST_SUCCESS = '[PROJECT API] List Success',
    LIST_CANCEL = '[PROJECT API] List Cancel',

    DETAILS = '[PROJECT API] Details',
    DETAILS_SUCCESS = '[PROJECT API] Details Success',
    DETAILS_CANCEL = '[PROJECT API] Details Cancel',

    CREATE = '[PROJECT API] Create',
    CREATE_SUCCESS = '[PROJECT API] Create Success',
    CREATE_CANCEL = '[PROJECT API] Create Cancel',

    UPDATE = '[PROJECT API] Update',
    UPDATE_SUCCESS = '[PROJECT API] Update Success',
    UPDATE_CANCEL = '[PROJECT API] Update Cancel',

    DELETE = '[PROJECT API] Delete',
    DELETE_SUCCESS = '[PROJECT API] Delete Success',
    DELETE_CANCEL = '[PROJECT API] Delete Cancel'
}